fix, RAM: the module now runs if its config is empty
Before this change there was a problem that if a module is started for the
first time, its configuration is set (namely `"enabled": true`, but the
internal state does not change, and since the module is disabled by default
`m_enabled == false`, it remains disabled. This change fixes this, since now
the module status information is loaded from the config in any case, even if
a new config has just been created.