fix inability to switch tabs with mods
Reworked the way tabs are stored, now they are no longer newly created in `Renderer`
which used to be passed to `m_screen.Loop`, but stored in a class. When necessary,
their data is updated in the `set_data` method. Also fixed a major bug in `Tabs::OnRender()`
which was that a reference to m_tabs[m_current_tab] was taken, but if m_tabs was empty
it would result in a UB. The problem was discovered when building in Debug mode