feat: add lazy module initialization
Prior to this, I had to initialize modules every time smu-server was started, even if it was
started with the `--help` or `--version` key (this does not involve a full server startup).
This resulted in additional overhead and unnecessary log output if an error occurred when
registering a module. Now modules are added to a special queue, and they are registered
only when `Application::run()` is called.