fix text truncation about connection error if terminal length is insufficient
The error occurred because the text was displayed in Raw mode (before the UI
stopped, which switches the terminal to normal mode). This issue has been
resolved, the code has been refactored, and a line about the error to be
displayed upon exit (optional) has been added to `Application::exit`.